# Backend/swagger.yml
openapi: 3.0.0
info:
  title: "KEV API"
  description: "An API made by KEV"
  version: "1.0.0"
servers:
  [url: "/api"]
...
paths:
...
  /login:
    get:
      operationId: "auth.verify_password"
      tags: ["Auth"]
      summary: "Authentication using username/password OR token"
      parameters:
        - name: username_or_token
          in: path
          description: username or token to be used for authentication
          required: true
          schema:
            type: string
        - name: password
          in: path
          description: password to be used for authentication, can be left blank if using token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Successfully authenticated"